LayoutTests/imported/w3c/web-platform-tests/websockets/bufferedAmount-unchanged-by-sync-xhr.any.worker-expected.txt


FAIL bufferedAmount should not be updated during a sync XHR assert_equals: expected 5 but got 0
